# contracts/corpus-ingestion/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Norvii Corpus Ingestion API",
    "version": "1.0.0"
  },
  "servers": [{"url": "/api/v1"}],
  "paths": {
    "/corpora": {
      "get": {"operationId": "listCorpora", "responses": {"200": {"description": "Corpus list", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Corpus"}}}}}}},
      "post": {"operationId": "createCorpus", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CorpusWrite"}}}}, "responses": {"201": {"description": "Created corpus", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Corpus"}}}}, "400": {"$ref": "#/components/responses/Error"}}}
    },
    "/corpora/{corpusId}": {
      "parameters": [{"$ref": "#/components/parameters/CorpusId"}],
      "get": {"operationId": "getCorpus", "responses": {"200": {"description": "Corpus", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Corpus"}}}}, "404": {"$ref": "#/components/responses/Error"}}},
      "patch": {"operationId": "updateCorpus", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CorpusUpdate"}}}}, "responses": {"200": {"description": "Updated corpus", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Corpus"}}}}, "409": {"$ref": "#/components/responses/Error"}}}
    },
    "/corpora/{corpusId}/disable": {"post": {"operationId": "disableCorpus", "parameters": [{"$ref": "#/components/parameters/CorpusId"}], "requestBody": {"$ref": "#/components/requestBodies/VersionCommand"}, "responses": {"200": {"description": "Disabled corpus", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Corpus"}}}}, "409": {"$ref": "#/components/responses/Error"}}}},
    "/corpora/{corpusId}/enable": {"post": {"operationId": "enableCorpus", "parameters": [{"$ref": "#/components/parameters/CorpusId"}], "requestBody": {"$ref": "#/components/requestBodies/VersionCommand"}, "responses": {"200": {"description": "Enabled corpus", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Corpus"}}}}, "409": {"$ref": "#/components/responses/Error"}}}},
...
    "/corpora/{corpusId}/sources/{sourceId}/document": {"get": {"operationId": "getLatestDocument", "parameters": [{"$ref": "#/components/parameters/CorpusId"}, {"$ref": "#/components/parameters/SourceId"}], "responses": {"200": {"description": "Latest ready document", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Document"}}}}, "404": {"$ref": "#/components/responses/Error"}}}},
...
  },
  "components": {
    "parameters": {
      "CorpusId": {"name": "corpusId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "SourceId": {"name": "sourceId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
    },
    "requestBodies": {"VersionCommand": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["version"], "properties": {"version": {"type": "integer", "minimum": 1}}}}}}},
    "responses": {"Error": {"description": "Safe error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}},
    "schemas": {
      "CorpusWrite": {"type": "object", "additionalProperties": false, "required": ["name", "description", "language", "jurisdiction"], "properties": {"name": {"type": "string", "minLength": 1}, "description": {"type": "string", "minLength": 1}, "language": {"type": "string", "enum": ["en", "pt"]}, "jurisdiction": {"type": "string", "minLength": 1}}},
      "CorpusUpdate": {"allOf": [{"$ref": "#/components/schemas/CorpusWrite"}, {"type": "object", "required": ["version"], "properties": {"version": {"type": "integer", "minimum": 1}}}]},
      "Corpus": {"allOf": [{"$ref": "#/components/schemas/CorpusWrite"}, {"type": "object", "required": ["id", "status", "sourceCount", "version", "createdAt", "updatedAt"], "properties": {"id": {"type": "string", "format": "uuid"}, "status": {"type": "string", "enum": ["enabled", "disabled"]}, "sourceCount": {"type": "integer", "minimum": 0}, "version": {"type": "integer", "minimum": 1}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}}}]},
...
      "DocumentUnit": {"type": "object", "required": ["id", "kind", "ordinal", "startOffset", "endOffset", "locator", "contentSha256"], "properties": {"id": {"type": "string", "format": "uuid"}, "parentId": {"type": ["string", "null"], "format": "uuid"}, "kind": {"type": "string", "enum": ["document", "title", "chapter", "section", "article", "paragraph", "item", "recital", "page", "block"]}, "ordinal": {"type": "integer", "minimum": 0}, "marker": {"type": ["string", "null"]}, "label": {"type": ["string", "null"]}, "startOffset": {"type": "integer", "minimum": 0}, "endOffset": {"type": "integer", "minimum": 0}, "locator": {"type": "string"}, "contentSha256": {"type": "string", "pattern": "^[0-9a-f]{64}$"}}},
      "Document": {"type": "object", "required": ["id", "sourceRevisionId", "pipelineVersion", "text", "textSha256", "createdAt", "units", "provenance"], "properties": {"id": {"type": "string", "format": "uuid"}, "sourceRevisionId": {"type": "string", "format": "uuid"}, "pipelineVersion": {"type": "string"}, "text": {"type": "string"}, "textSha256": {"type": "string", "pattern": "^[0-9a-f]{64}$"}, "createdAt": {"type": "string", "format": "date-time"}, "units": {"type": "array", "items": {"$ref": "#/components/schemas/DocumentUnit"}}, "provenance": {"$ref": "#/components/schemas/RevisionProvenance"}}},
      "RevisionProvenance": {"type": "object", "required": ["contentSha256", "capturedAt", "mediaType", "byteSize", "finalUrl", "extractedContentSha256"], "properties": {"contentSha256": {"type": "string", "pattern": "^[0-9a-f]{64}$"}, "capturedAt": {"type": "string", "format": "date-time"}, "mediaType": {"type": "string"}, "byteSize": {"type": "integer", "minimum": 1}, "finalUrl": {"type": ["string", "null"], "format": "uri"}, "extractedContentSha256": {"type": ["string", "null"], "pattern": "^[0-9a-f]{64}$"}}},
      "ErrorEnvelope": {"type": "object", "required": ["error"], "properties": {"error": {"type": "object", "required": ["code", "message", "requestId"], "properties": {"code": {"type": "string", "enum": ["invalid_input", "payload_too_large", "unsafe_url", "unsupported_content", "duplicate_source", "stale_state", "not_found", "unavailable", "acquisition_failed", "extraction_failed", "publication_failed", "internal_error"]}, "message": {"type": "string"}, "fields": {"type": "object", "additionalProperties": {"type": "string"}}, "requestId": {"type": "string", "format": "uuid"}}}}
    }
  }
}
}